Added creds-fetch API to 'cray scsd bmc certs'.
Added a new API endpoint to SCSD to allow for the fetching of
BMC creds from Vault.  This is to be used primarily by services
within the service mesh; thus it is important that the OPA
filtering rules strictly preclude any other entry into this
API.

* Resolves CASMHMS-5273

Tested on:

* local build machine

Ran a copy of the new SCSD on my build machine.  Ran the modified CLI
locally as well according to the CLI development guidelines.  Everything
worked as expected.

Low risk.