PE-40340 Deprecate PALS commands
Recently, we've refactored PALS on Shasta systems to avoid using the
REST API, so it works like it does on HPCM systems. The cray pals,
aprun, and mpiexec commands fail on updated systems.

To handle this, mark the commands as hidden so they don't show up in the
help message, and add a message to aprun/mpiexec to assist with the
transition.

Tested manually on slice.