CASMCMS-8192: Remove version file
The .version file determines the version of the rpm.
The version of the RPM should be the same as the CLI, so I am using
the .version file in two places, the RPM spec file and the setup.py
file. Therefore, I could eliminate the 'version' file (without the
dot).