CASMCMS-8192: Default BOS to v1 (#59)
* CASMCMS-8192: Default BOS to v1 and remove version file

When the Cray CLI is used with the `bos` command and no version string
is specified, BOS will default to using the `v1` version.

The .version file determines the version of the rpm.
The version of the RPM should be the same as the CLI, so I am using
the .version file in two places, the RPM spec file and the setup.py
file. Therefore, I could eliminate the 'version' file (without the
dot).