Newer `urllib` handling
The `urllib.parse.urlsplit()` in Python 3.9 and newer does like
scheme-less URLs that contain ports. If a URL is passed without a scheme
but has a port, then the entire string before the port is identified as
the scheme.

e.g. `example.local:443` sets `scheme` to `example.local`, `path` to
443, and `port` to `None`.

By appended `//` to the beginning of a URL without a scheme,
`urllib.parse.urlsplit()` properly identifies each lexicon.