Update Development docs
In the previous refactor (#81) the development documentation and scripts
needed more attention.

This updates the `Developing.md` with up-to-date information, as well as
updating the `utils/` items.

The `devenv` container is no longer used, instead the virtualenv path is
now recommended.

The swagger container was updated and pushed to our Artifactory on
algol60.net.

This also updates the `--version` and `--help` (`-h`) arguments to
leverage `click`'s builtins, removing the funky `build_version` handling
(which doesn't work anyway given we use `setuptools_scm[toml]` now).

`Makefile` was updated for maintainers to make and publish new swagger
Docker images as needed.

Add setup-python directions